+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed error when loading DOMAIN object with NOT NULL constraint.
 - Fixed a bug in generating code for a DOMAIN object with constraints.
+- Fixed UI freeze when testing connection with database.
 
 ### Removed